Update with new spec
