+++ packages/jest-core/src/TestScheduler.ts
@@ -12,7 +12,6 @@ import {
   CoverageReporter,
   DefaultReporter,
   GitHubActionsReporter,
-  GithubActionsLogsReporter,
   BaseReporter as JestReporter,
   NotifyReporter,
   Reporter,
@@ -343,15 +342,9 @@ class TestScheduler {
       switch (reporter) {
         case 'default':
           summary = true;
-          if (verbose) {
-            this.addReporter(new VerboseReporter(this._globalConfig));
-          } else {
-            GITHUB_ACTIONS
-              ? this.addReporter(
-                  new GithubActionsLogsReporter(this._globalConfig),
-                )
-              : this.addReporter(new DefaultReporter(this._globalConfig));
-          }
+          verbose
+            ? this.addReporter(new VerboseReporter(this._globalConfig))
+            : this.addReporter(new DefaultReporter(this._globalConfig));
           break;
         case 'github-actions':
           GITHUB_ACTIONS && this.addReporter(new GitHubActionsReporter());
