+++ packages/jest-reporters/src/index.ts
@@ -30,7 +30,6 @@ export {default as GitHubActionsReporter} from './GitHubActionsReporter';
 export {default as NotifyReporter} from './NotifyReporter';
 export {default as SummaryReporter} from './SummaryReporter';
 export {default as VerboseReporter} from './VerboseReporter';
-export {default as GithubActionsLogsReporter} from './GithubActionsLogsReporter';
 export type {
   Reporter,
   ReporterOnStartOptions,
